helper/cloud_adsperformance_creative_mapping/requirements.txt
google-cloud-bigquery==3.15.0
google-auth==2.25.2
pandas==2.1.4
pytz==2023.3
pyarrow==14.0.1
requests>=2.31.0
